+++ common/src/main/java/net/qbar/common/grid/IConnectionAware.java
@@ -1,6 +1,7 @@
 package net.qbar.common.grid;
 
 import net.minecraft.util.EnumFacing;
+import net.qbar.common.grid.impl.CableGrid;
 
 public interface IConnectionAware
 {
