--- common/src/main/java/net/qbar/common/grid/IFluidPipe.java
@@ -1,6 +1,0 @@
-package net.qbar.common.grid;
-
-public interface IFluidPipe extends ITileCable<PipeGrid>
-{
-    void fillNeighbors();
-}
