+++ common/src/main/java/net/qbar/common/grid/node/ISteamPipe.java
@@ -1,5 +1,7 @@
-package net.qbar.common.grid;
+package net.qbar.common.grid.node;
 
+import net.qbar.common.grid.impl.SteamGrid;
+import net.qbar.common.grid.node.ITileCable;
 import net.qbar.common.steam.ISteamHandler;
 
 import java.util.Collection;
