+++ common/src/main/java/net/qbar/common/multiblock/TileMultiblockGag.java
@@ -4,7 +4,7 @@ import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.common.capabilities.Capability;
-import net.qbar.common.grid.CableGrid;
+import net.qbar.common.grid.impl.CableGrid;
 import net.qbar.common.grid.IConnectionAware;
 import net.qbar.common.tile.ITileInfoProvider;
 import net.qbar.common.tile.QBarTileBase;
