+++ common/src/test/java/net/qbar/common/grid/ITileCableTestImpl.java
@@ -3,6 +3,9 @@ package net.qbar.common.grid;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.qbar.common.grid.impl.CableGrid;
+import net.qbar.common.grid.node.ITileCable;
+import net.qbar.common.grid.node.ITileNode;
 
 import java.util.EnumMap;
