+++ common/src/test/java/net/qbar/common/grid/PipeGridTest.java
@@ -1,5 +1,7 @@
 package net.qbar.common.grid;
 
+import net.qbar.common.grid.impl.PipeGrid;
+import net.qbar.common.grid.node.IFluidPipe;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
