+++ content/logistic/src/main/java/net/qbar/common/tile/TileSteamPipe.java
@@ -5,9 +5,9 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.common.capabilities.Capability;
-import net.qbar.common.grid.ISteamPipe;
-import net.qbar.common.grid.ITileNode;
-import net.qbar.common.grid.SteamGrid;
+import net.qbar.common.grid.node.ISteamPipe;
+import net.qbar.common.grid.node.ITileNode;
+import net.qbar.common.grid.impl.SteamGrid;
 import net.qbar.common.steam.CapabilitySteamHandler;
 import net.qbar.common.steam.ISteamHandler;
 import net.qbar.common.steam.SteamUtil;
