+++ content/logistic/src/main/java/net/qbar/common/tile/TileSteamValve.java
@@ -6,9 +6,9 @@ import net.minecraft.util.EnumFacing;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.fluids.FluidStack;
 import net.qbar.common.block.BlockSteamValve;
-import net.qbar.common.grid.CableGrid;
+import net.qbar.common.grid.impl.CableGrid;
 import net.qbar.common.grid.IConnectionAware;
-import net.qbar.common.grid.SteamGrid;
+import net.qbar.common.grid.impl.SteamGrid;
 import net.qbar.common.steam.CapabilitySteamHandler;
 import net.qbar.common.steam.ISteamTank;
 import net.qbar.common.steam.SteamTank;
