+++ content/logistic/src/main/java/net/qbar/common/tile/machine/TileBelt.java
@@ -12,6 +12,11 @@ import net.qbar.client.render.model.obj.QBarOBJState;
 import net.qbar.common.block.BlockBelt.EBeltSlope;
 import net.qbar.common.event.TickHandler;
 import net.qbar.common.grid.*;
+import net.qbar.common.grid.impl.BeltGrid;
+import net.qbar.common.grid.impl.CableGrid;
+import net.qbar.common.grid.node.IBelt;
+import net.qbar.common.grid.node.ITileCable;
+import net.qbar.common.grid.node.ITileNode;
 import net.qbar.common.steam.CapabilitySteamHandler;
 import net.qbar.common.steam.ISteamHandler;
 import net.qbar.common.steam.SteamUtil;
