+++ content/machine/src/main/java/net/qbar/common/tile/machine/TileKeypunch.java
@@ -26,7 +26,7 @@ import net.qbar.common.container.BuiltContainer;
 import net.qbar.common.container.ContainerBuilder;
 import net.qbar.common.container.EmptyContainer;
 import net.qbar.common.event.TickHandler;
-import net.qbar.common.grid.ITileWorkshop;
+import net.qbar.common.grid.node.ITileWorkshop;
 import net.qbar.common.grid.WorkshopMachine;
 import net.qbar.common.gui.MachineGui;
 import net.qbar.common.init.QBarItems;
