+++ content/machine/src/main/java/net/qbar/common/tile/machine/TileTank.java
@@ -14,7 +14,7 @@ import net.qbar.common.container.BuiltContainer;
 import net.qbar.common.container.ContainerBuilder;
 import net.qbar.common.container.IContainerProvider;
 import net.qbar.common.fluid.DirectionalTank;
-import net.qbar.common.grid.CableGrid;
+import net.qbar.common.grid.impl.CableGrid;
 import net.qbar.common.grid.IConnectionAware;
 import net.qbar.common.gui.MachineGui;
 import net.qbar.common.init.QBarItems;
